- Show properties on single click instead of double click about #70 but do not open files in simple click
- About #62 fix reading of *.gap files (EcosTheta receiver results)
- Fix language selection dialog (was empty)
- About #61 Removes 'Average' value for Schroeder's curves .gabe files
- About #60 #12 Change number of decimal depending of value unit
- About #58 By default, display all elements of the tree for each calculation code in the 'Calculation tab'
- About #56 do not set readonly when directivity file is chosen, add description in directivity
- About #55 activate copy paste of directivity bdd node
- Remove two unused properties about #54
- About #51 Left align labels in the properties grid ui
- Set NaN for global values for reverberation time in classical theory. About #13 
- Double precision of preprocess program
- Add graphy lib used by report plugin
- Fix major memory issue with tetramesh read/write that may raise lost particle to 100%

+++ src/preprocess/computations.hpp
@@ -73,15 +73,15 @@ private:
 	
 	int FindFaceWithSommet( const ivec3& sommetsSearch );
 	//Retourne -1 si aucun point correspondant
-	int FindIndexWithPosition( const vec3& position );
+	int FindIndexWithPosition( const dvec3& position );
 	//return vrai si la division a eu lieu
 	//addedFaceOne -1 ou idface (si inséré)
 	//addedFaceTwo -1 ou idface (si inséré)
 	bool SplitTriangleByThree( formatPOLY::t_face& triangleToSplit , const int& splitCentre,int* addedFaceOne,int* addedFaceTwo  );
 
 	//return faux si aucun traitement effectué
-	bool OnCollisionDetectedSplitIt( formatPOLY::t_face& triangleOne, formatPOLY::t_face& triangleTwo, const vec3& position );
-	bool OnCollisionDetectedDestroyIt( formatPOLY::t_face& triangleOne, formatPOLY::t_face& triangleTwo, const vec3& position );
+	bool OnCollisionDetectedSplitIt( formatPOLY::t_face& triangleOne, formatPOLY::t_face& triangleTwo, const dvec3& position );
+	bool OnCollisionDetectedDestroyIt( formatPOLY::t_face& triangleOne, formatPOLY::t_face& triangleTwo, const dvec3& position );
 	void PushNewFace(const formatPOLY::t_face& newFace);
 	void ReBuildOctree();
 	bool coplanarIntersection(int idface,int idfaceTest);
